refactor: split dialog components into modules
