+++ frontends/nextjs/src/components/ui/organisms/dialogs/Command.tsx
@@ -1,3 +1,6 @@
 'use client'
 
 export * from './command'
+export * from './Command/Palette'
+export * from './Command/Results'
+export * from './Command/useCommandState'
